requirements.txt
python-dotenv
python-telegram-bot
requests
Flask
gunicorn
quart
fastapi
uvicorn
